requirements.txt
fastapi
uvicorn
gunicorn
openai
langchain
langchain-experimental
python-dotenv
langchain-openai